requirements.txt
torch
torchvision
Pillow
Flask
flask-cors
transformers
numpy
scikit-learn
python-dotenv
